* Refactored "Star Search Minigame" to watch for mouse clicks, and when a star is clicked to pass that information back to the Constellation class.
