--- Assets/HeartFollower.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 489577ba7f9c4e94c937f38ef3a02324
-timeCreated: 1529963603
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
