--- Assets/Prefabs/Constellation.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: f9f9ee2b6c8fab84695ee5376775b76f
-timeCreated: 1530138574
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
