+++ Assets/Prefabs/Constellations/Key.prefab
@@ -421,8 +421,8 @@ Transform:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1989795154521286}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -6.3, y: -1.5, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalPosition: {x: 0, y: -1.1, z: 0}
+  m_LocalScale: {x: 1.1500001, y: 1, z: 1}
   m_Children:
   - {fileID: 4036065114896050}
   - {fileID: 4525204078319840}
@@ -986,7 +986,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   MiniStar: {fileID: 1589996901234664, guid: 35c047f166bb8894f90e789e0e581e0f, type: 2}
-  smallStarGap: 0.05
+  smallStarGap: 0.08
 --- !u!114 &114114145663366066
 MonoBehaviour:
   m_ObjectHideFlags: 1
