+++ Assets/Scripts/CastleBGMEvents.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 469069d6339fad641ae0b8a8f71eb36f
+timeCreated: 1524903052
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
