+++ Assets/Scripts/StarSearchMinigame.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 7bbfd6375612d6249a585e634a05d5ef
+timeCreated: 1530657628
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
