--- Assets/StarSearchMinigame.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: df0919bd54852db408ef504f8efb1a53
-timeCreated: 1530050760
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
